google-review-craw/requirements.txt
seleniumbase>=4.34.9
googletrans==4.0.2
tqdm>=4.66.3
pymongo==4.12.0
boto3==1.35.1
fastapi==0.104.1
uvicorn==0.24.0
botocore~=1.35.99
pydantic~=2.11.5
requests>=2.31.0
PyYAML>=6.0
